Conflicts:
	.github/workflows/continuous_integration.yml
	README.md
	src/common/2d/v_draw.cpp
	src/common/2d/v_draw.h
	src/common/rendering/gl/gl_postprocess.cpp
	src/common/rendering/gles/gles_framebuffer.h
	src/common/rendering/hwrenderer/data/hw_vrmodes.cpp
	src/common/rendering/hwrenderer/data/hw_vrmodes.h
	src/common/rendering/hwrenderer/hw_draw2d.cpp
	src/d_main.cpp
	src/g_statusbar/sbar_mugshot.cpp
	src/playsim/actor.h
	src/playsim/p_map.cpp
	src/playsim/p_mobj.cpp
	src/r_data/models.cpp
	src/rendering/2d/v_blend.cpp
	src/rendering/hwrenderer/hw_entrypoint.cpp
	src/rendering/hwrenderer/scene/hw_drawinfo.cpp
	src/rendering/hwrenderer/scene/hw_sprites.cpp
	src/rendering/hwrenderer/scene/hw_weapon.cpp
	src/rendering/hwrenderer/scene/hw_weapon.h
	src/rendering/r_utility.cpp
	src/rendering/r_utility.h
	src/version.h
	wadsrc/static/language.0
	wadsrc/static/language.csv
	wadsrc/static/menudef.txt
	wadsrc/static/zscript/actors/inventory/stateprovider.zs

+++ src/common/utility/zstring.cpp
@@ -37,6 +37,7 @@
 #include <string.h>
 #include <new>		// for bad_alloc
 
+
 #include "zstring.h"
 #include "utf8.h"
 #include "stb_sprintf.h"
@@ -603,7 +604,7 @@ ptrdiff_t FString::LastIndexOfAny (const FString &charset) const
 
 ptrdiff_t FString::LastIndexOfAny (const char *charset) const
 {
-	return LastIndexOfAny (charset, long(Len()));
+	return LastIndexOfAny (charset, ptrdiff_t(Len()));
 }
 
 ptrdiff_t FString::LastIndexOfAny (const FString &charset, ptrdiff_t endIndex) const
